+++ examples/ssr/lib/home_page.dart
@@ -1,4 +1,4 @@
-import 'package:react_router/react_router.dart';
+import 'package:react_router_dom/react_router_dom.dart';
 import 'package:react_web/react_web.dart' hide link;
 
 import 'hooks_page.dart' show memoizedCounter;